♻️ move demo data to components directory for better code organization
